+++ src/main/java/NotModified304/Scatch/dto/timeTable/response/TimeTableDetailResponse.java
@@ -1,6 +1,6 @@
 package NotModified304.Scatch.dto.timeTable.response;
 
-import NotModified304.Scatch.domain.todo.TimeTableDetail;
+import NotModified304.Scatch.domain.timeTable.TimeTableDetail;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
